Create OpenEntryQueue component.
